requirements.txt
Django>= 5.0, < 6.0
djangorestframework
django_filter>=23.5.0
djangorestframework-gis==1.0
django-cors-headers==4.3.1
pyvips==2.2.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-admin-interface==0.28.3
drf-generators==0.5.0
drf-dynamic-fields==0.4.0
django-cleanup==8.0.0
django-polymorphic>=3.1.0
django-admin-autocomplete-filter==0.7.1
django-admin-numeric-filter==0.1.9
django-leaflet-admin-list==0.0.7
django-ckeditor==6.7.0
django-admin-rangefilter==0.12.0
django-markdownfield
djangorestframework-xml==2.0.0